+++ src/main/java/org/example/Sec3/Flux/FluxFromStream.java
@@ -1,4 +1,4 @@
-package org.example.Flux.sec1;
+package org.example.Sec3.Flux;
 
 import java.util.List;
 import java.util.stream.Stream;
