+++ src/main/java/org/example/Sec3/Flux/FluxInterval.java
@@ -1,4 +1,4 @@
-package org.example.Flux.sec1;
+package org.example.Sec3.Flux;
 
 import java.time.Duration;
