+++ src/main/java/org/example/Sec3/Flux/FluxJust.java
@@ -1,4 +1,4 @@
-package org.example.Flux.sec1;
+package org.example.Sec3.Flux;
 
 import org.example.commons.Util;
 import reactor.core.publisher.Flux;
